@@ -1,34 +1,55 @@
 source 'https://rubygems.org'
+ruby '1.9.3'
+#ruby-gemset=railstutorial_rails_4_0
 
-gem 'rails', '3.2.13'
-gem 'bootstrap-sass', '2.1'
+gem 'rails', '4.0.2'
+gem 'bootstrap-sass', '2.3.2.0'
+gem 'bcrypt-ruby', '3.1.2'
+gem 'faker', '1.1.2'
+gem 'will_paginate', '3.0.4'
+gem 'bootstrap-will_paginate', '0.0.9'
 
 group :development, :test do
-  gem 'sqlite3', '1.3.5'
-  gem 'rspec-rails', '2.11.0'
-  gem 'guard-rspec', '1.2.1'
-
-    gem 'guard-spork', '1.2.0'
-  gem 'spork', '0.9.2'
+  gem 'sqlite3', '1.3.8'
+  gem 'rspec-rails', '2.13.1'
+  # The following optional lines are part of the advanced setup.
+  gem 'guard-rspec', '2.5.0'
+  gem 'spork-rails', '4.0.0'
+  gem 'guard-spork', '1.5.0'
+  gem 'childprocess', '0.3.6'
 end
 
-# Gems used only for assets and not required
-# in production environments by default.
-group :assets do
-  gem 'sass-rails',   '3.2.5'
-  gem 'coffee-rails', '3.2.2'
-  gem 'uglifier', '1.2.3'
-end
-
-gem 'jquery-rails', '2.0.2'
-
-# Test gems on Macintosh OS X
 group :test do
-  gem 'capybara', '1.1.2'
-  gem 'rb-fsevent', '0.9.1', :require => false
+  gem 'selenium-webdriver', '2.35.1'
+  gem 'capybara', '2.1.0'
+  gem 'factory_girl_rails', '4.2.0'
+  gem 'cucumber-rails', '1.3.0', :require => false
+  gem 'database_cleaner', github: 'bmabey/database_cleaner'
+
+  # Uncomment this line on OS X.
   gem 'growl', '1.0.3'
-end 
+
+  # Uncomment these lines on Linux.
+  # gem 'libnotify', '0.8.0'
+
+  # Uncomment these lines on Windows.
+  # gem 'rb-notifu', '0.0.4'
+  # gem 'win32console', '1.3.2'
+  # gem 'wdm', '0.1.0'
+end
+
+gem 'sass-rails', '4.0.1'
+gem 'uglifier', '2.1.1'
+gem 'coffee-rails', '4.0.1'
+gem 'jquery-rails', '3.0.4'
+gem 'turbolinks', '1.1.1'
+gem 'jbuilder', '1.0.2'
+
+group :doc do
+  gem 'sdoc', '0.3.20', require: false
+end
 
 group :production do
-  gem 'pg', '0.12.2'
+  gem 'pg', '0.15.1'
+  gem 'rails_12factor', '0.0.2'
 end
