@@ -2,34 +2,34 @@ require 'spec_helper'
 
 describe "Static pages" do
 
-  subject { page }
+    subject { page }
 
   describe "Home page" do
     before { visit root_path }
 
-    it { should have_selector('h1',    text: 'Sample App') }
-    it { should have_selector('title', text: full_title('')) }
-    it { should_not have_selector 'title', text: '| Home' }
+    it { should have_content('Sample App') }
+    it { should have_title(full_title('')) }
+    #it { should_not have_title('| Home') }
   end
 
   describe "Help page" do
     before { visit help_path }
 
-    it { should have_selector('h1',    text: 'Help') }
-    it { should have_selector('title', text: full_title('Help')) }
+    it { should have_content('Help') }
+    it { should have_title(full_title('Help')) }
   end
 
   describe "About page" do
     before { visit about_path }
 
-    it { should have_selector('h1',    text: 'About') }
-    it { should have_selector('title', text: full_title('About Us')) }
+    it { should have_content('About') }
+    it { should have_title(full_title('About Us')) }
   end
 
-  describe "Contact page" do
+    describe "Contact page" do
     before { visit contact_path }
 
-    it { should have_selector('h1',    text: 'Contact') }
-    it { should have_selector('title', text: full_title('Contact')) }
+    it { should have_selector('h1', text: 'Contact') }
+    it { should have_title(full_title('Contact')) }
   end
 end
