@@ -7,7 +7,7 @@ describe "User pages" do
   describe "signup page" do
     before { visit signup_path }
 
-    it { should have_selector('h1',    text: 'Sign up') }
-    it { should have_selector('title', text: full_title('Sign up')) }
+    it { should have_content('Sign up') }
+    it { should have_title(full_title('Sign up')) }
   end
 end
